Add support for specifying the a checkout ref
This is useful in cases where a repository is importing itself as
a private repo in a workflow running for a pull request. In these
cases, it is be necessary to pass a custom ref to the
action/checkout action so that it checks out the private repo in
the correct manner instead of using the PR ref.